refactor test namespace
