+++ test/Device/DeviceDesktopTest.cs
@@ -3,7 +3,7 @@
 
 using Xunit;
 
-namespace Wangkanai.Detection.Test
+namespace Wangkanai.Detection.DeviceTest
 {
     public class DeviceDesktopTest : DeviceTestAbstract
     {
