+++ test/Device/DeviceMobileTest.cs
@@ -3,7 +3,7 @@
 
 using Xunit;
 
-namespace Wangkanai.Detection.Test
+namespace Wangkanai.Detection.DeviceTest
 {
     public class DeviceMobileTest : DeviceTestAbstract
     {
