+++ test/Device/DeviceTabletTest.cs
@@ -1,9 +1,10 @@
 // Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
+using Wangkanai.Detection.Test;
 using Xunit;
 
-namespace Wangkanai.Detection.Test
+namespace Wangkanai.Detection.DeviceTest
 {
     public class DeviceTabletTest : DeviceTestAbstract
     {
