+++ test/Device/DeviceTest.cs
@@ -5,7 +5,7 @@ using System;
 
 using Xunit;
 
-namespace Wangkanai.Detection.Test
+namespace Wangkanai.Detection.DeviceTest
 {
     public class DeviceTest
     {
