+++ test/Device/DeviceTestAbstract.cs
@@ -7,7 +7,7 @@ using Moq;
 
 using Wangkanai.Detection.Services;
 
-namespace Wangkanai.Detection.Test
+namespace Wangkanai.Detection.DeviceTest
 {
     public abstract class DeviceTestAbstract
     {
