+++ test/Engine/EngineTest.cs
@@ -3,7 +3,7 @@
 
 using Xunit;
 
-namespace Wangkanai.Detection.Test
+namespace Wangkanai.Detection.EngineTest
 {
     public class EngineTest
     {
