+++ test/Models/VersionTests.cs
@@ -5,7 +5,7 @@ using System;
 
 using Xunit;
 
-namespace Wangkanai.Detection.Test
+namespace Wangkanai.Detection.Models
 {
     public class VersionTests
     {
